// frontend/frontend/app/src/app/model/subjectModel.ts
export class SubjectModel{

    name: string;
    typeOfSubject: string;
    yearOfStudies: number;
    department: string;
    subjectCode: string;
    classesFond: number;
    numberOfPoints: number;
    goalOfSubject: string;
    outcomeOfSubject:string;
    excerciseTerm: string;
    lectureTerm: string;
    propositions: string;
    laboratory: string;
    semester: number;
    id:number;
    master: number;
    v:string;
    p:string;
    l:string;
    eq:string;
    comment: string;
    urleq:string;
    urlv:string;
    urlp:string;
    urll:string;

}